+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2024-02-25]
+## [1.1.0 - 2024-02-26]
 ### Added
 - Dockerfile
 - Single Spa
@@ -32,3 +32,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Download authentication script
 - Exists Votum and Application Form
 - Exists Script
+- Fetch publications and other documents